remove MCPConfigurationServerRunner altogether
