+++ front/temporal/agent_loop/activities/run_model.ts
@@ -5,7 +5,7 @@ import {
   tryListMCPTools,
 } from "@app/lib/actions/mcp_actions";
 import type { InternalMCPServerNameType } from "@app/lib/actions/mcp_internal_actions/constants";
-import { buildToolSpecification } from "@app/lib/actions/runners";
+import { buildToolSpecification } from "@app/lib/actions/mcp";
 import {
   isDustAppChatBlockType,
   runActionStreamed,
